refactor: filter config 분리
