+++ src/main/java/team_questio/questio/security/config/SecurityConfig.java
@@ -3,32 +3,22 @@ package team_questio.questio.security.config;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.filter.CorsFilter;
 import team_questio.questio.security.filter.JWTAuthenticationFilter;
-import team_questio.questio.security.util.JWTTokenProvider;
 
 @Configuration
 @RequiredArgsConstructor
 public class SecurityConfig {
-    private final SecurityProperties securityProperties;
-
-    @Bean
-    public BCryptPasswordEncoder bCryptPasswordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
+    private final CorsFilter corsFilter;
+    private final JWTAuthenticationFilter jwtAuthenticationFilter;
 
     @Bean
     public SecurityFilterChain filterChain(
-            HttpSecurity http,
-            CorsFilter corsFilter,
-            UsernamePasswordAuthenticationFilter usernamePasswordAuthenticationFilter
+            HttpSecurity http
     ) throws Exception {
         http.csrf(AbstractHttpConfigurer::disable);
         http.httpBasic(AbstractHttpConfigurer::disable);
@@ -43,25 +33,13 @@ public class SecurityConfig {
         );
 
         http.addFilter(corsFilter);
-        http.addFilter(usernamePasswordAuthenticationFilter);
+        http.addFilter(jwtAuthenticationFilter);
 
         return http.build();
     }
 
     @Bean
-    public UsernamePasswordAuthenticationFilter usernamePasswordAuthenticationFilter(
-            AuthenticationManager authenticationManager,
-            JWTTokenProvider jwtTokenProvider
-    ) {
-        var filter = new JWTAuthenticationFilter(authenticationManager, jwtTokenProvider);
-        filter.setFilterProcessesUrl(securityProperties.loginUrl());
-
-        return filter;
-    }
-
-    @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration)
-            throws Exception {
-        return authenticationConfiguration.getAuthenticationManager();
+    public BCryptPasswordEncoder bCryptPasswordEncoder() {
+        return new BCryptPasswordEncoder();
     }
 }
